+++ app/views/marketplaces/search-marketplace.js
@@ -1,5 +1,0 @@
-// Admin-Only View (gets overridden for admin)
-// Exists here so it can be rendered as an empty view
-import Ember from "ember";
-var SearchMarketplaceView = Ember.View.extend({});
-export default SearchMarketplaceView;