+++ app/views/results/search-results-dropdown-filter.js
@@ -1,0 +1,8 @@
+import ResultsDropdownFilterView from "./results-dropdown-filter";
+import { defineFilter } from "./results-dropdown-filter";
+
+var SearchResultsDropdownFilterView = ResultsDropdownFilterView.extend({
+	templateName: "results/search-results-dropdown-filter",
+});
+
+export default SearchResultsDropdownFilterView;