+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Balanced Dashboard Changelog
 
+### master
+* Updated search results view (#1489)
+* Added keyboard shortcuts to open and close the search modal
+* Bug fix for log search results
+* Display the most appropriate search results tab by default
+
 ### 1.1.2
 
 * Removed the orders messages from old marketplaces
+
+### 1.1.1
+
+* Bug fixes for order views
 
 ### 1.1.0